requirements.txt
telegram
pip
setuptools
mongoengine
pymongo
mongomock
google
google_auth_oauthlib
google-api-python-client
oauth2client
python-dateutil
python-telegram-bot==13.3
rstr
fuzzywuzzy
apscheduler
vobject
pyyaml
python-gitlab
requests[security]